Moved `Base.Chord.Note` to `Base.Core.Note` (notes are useful to more than chords so this is more logical)
